tweak: parser test restructure
