+++ src/lexer/mod.rs
@@ -1,5 +1,5 @@
-pub mod token;
 mod test;
+pub mod token;
 
 use crate::lexer::token::create_token;
 use token::Token;
